「✨」 feat(Structure): Moved push and remote to external file
