+++ src/pogit/gitRepo.py
@@ -6,7 +6,7 @@
 #    By: adjoly <adjoly@student.42angouleme.fr>     +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/12/05 17:10:56 by adjoly            #+#    #+#              #
-#    Updated: 2024/12/05 17:24:48 by adjoly           ###   ########.fr        #
+#    Updated: 2024/12/05 17:36:35 by adjoly           ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
